feat: consolidating channels
